CSP-81 Moved GetStandardDetails handler
