+++ src/SFA.DAS.Roatp.CourseManagement.Application/ProviderStandards/Queries/GetStandardDetails/GetStandardDetailsQueryResult.cs
@@ -1,6 +1,6 @@
 ﻿using SFA.DAS.Roatp.CourseManagement.Domain.ApiModels;
 
-namespace SFA.DAS.Roatp.CourseManagement.Application.Standard.Queries
+namespace SFA.DAS.Roatp.CourseManagement.Application.ProviderStandards.Queries.GetStandardDetails
 {
     public class GetStandardDetailsQueryResult
     {
